backend/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==4.0.1
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.136.1
greenlet==3.5.0
h11==0.16.0
httptools==0.7.1
idna==3.13
Mako==1.3.12
MarkupSafe==3.0.3
passlib==1.7.4
psycopg2-binary==2.9.12
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
PyJWT==2.12.1
python-dotenv==1.2.2
python-multipart==0.0.27
PyYAML==6.0.3
requests==2.33.1
resend==2.30.0
SQLAlchemy==2.0.49
starlette==1.0.0
tomli==2.4.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.46.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
